+++ src/features/mypage/lib/text.ts
@@ -1,0 +1,4 @@
+export const normalize = (value: unknown) =>
+  String(value ?? '')
+    .trim()
+    .toLowerCase()
