+++ src/features/mypage/ui/MyPage.tsx
@@ -1,413 +1,45 @@
 'use client'
 
-import {
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-  useSyncExternalStore,
-} from 'react'
-import { toast } from 'sonner'
-
-import PostFilter, { type SortOption } from '@/features/mypage/ui/PostFilter'
+import PostFilter from '@/features/mypage/ui/PostFilter'
 import MyPost from '@/features/mypage/ui/MyPost'
 import MyComment from '@/features/mypage/ui/MyComment'
 import MyLike from '@/features/mypage/ui/MyLike'
 import MyPageActionMenu from '@/features/mypage/ui/MyPageActionMenu'
 import { Pagination } from '@/shared/ui/pagination-je'
 import Profile from '@/features/mypage/ui/Profile'
-import TimeLine, { type TimelineItem } from '@/features/mypage/ui/TimeLine'
-import type {
-  MyCommentItem,
-  MyPagePostItem,
-} from '@/entities/mypage/model/mypage-ui-types'
-import { useSessionStore } from '@/entities/session/store/session-store'
-import { useMyPosts } from '@/features/mypage/hook/useMyPost'
-import { useMyComments } from '@/features/mypage/hook/useMyComments'
-import { useLikedPosts } from '@/features/mypage/hook/useLikes'
-import { useDeleteMyPosts } from '@/features/mypage/hook/useDeleteMyPost'
-import { useDeleteMyComments } from '@/features/mypage/hook/useDeleteMyComments'
-import { useTimelineHistory } from '@/features/mypage/hook/useTimeline'
+import TimeLine from '@/features/mypage/ui/TimeLine'
 
-type TabType = 'post' | 'comment' | 'like'
-
-const normalize = (v: unknown) =>
-  String(v ?? '')
-    .trim()
-    .toLowerCase()
-
-const formatDateParts = (input: string): { date: string; time: string } => {
-  const dateObj = new Date(input)
-  if (Number.isNaN(dateObj.getTime())) return { date: '', time: '' }
-  const yyyy = dateObj.getFullYear()
-  const mm = String(dateObj.getMonth() + 1).padStart(2, '0')
-  const dd = String(dateObj.getDate()).padStart(2, '0')
-  const hh = String(dateObj.getHours()).padStart(2, '0')
-  const min = String(dateObj.getMinutes()).padStart(2, '0')
-  return { date: `${yyyy}.${mm}.${dd}`, time: `${hh}:${min}` }
-}
-
-const getKstNow = () => {
-  const now = new Date()
-  const utc = now.getTime() + now.getTimezoneOffset() * 60_000
-  return new Date(utc + 9 * 60 * 60_000)
-}
-
-const startOfKstDay = () => {
-  const d = getKstNow()
-  d.setHours(0, 0, 0, 0)
-  return d
-}
-
-const addDays = (base: Date, offset: number) => {
-  const d = new Date(base)
-  d.setDate(base.getDate() + offset)
-  return d
-}
-
-const toYmd = (d: Date) => {
-  const yyyy = d.getFullYear()
-  const mm = String(d.getMonth() + 1).padStart(2, '0')
-  const dd = String(d.getDate()).padStart(2, '0')
-  return `${yyyy}-${mm}-${dd}`
-}
-
-const toMmDd = (d: Date) => {
-  const mm = String(d.getMonth() + 1).padStart(2, '0')
-  const dd = String(d.getDate()).padStart(2, '0')
-  return `${mm}.${dd}`
-}
-
-const toKoreanDay = (d: Date) => {
-  const days = ['일', '월', '화', '수', '목', '금', '토'] as const
-  return days[d.getDay()]
-}
-
-const DEFAULT_THUMBNAIL = '/images/mypage/post-example.png'
-const DEFAULT_AVATAR = '/images/profiles/default-1.webp'
+import { useMyPageController } from '@/features/mypage/hook/useMyPageController'
 
 const MyPage = () => {
-  const [tab, setTab] = useState<TabType>('post')
-  const [page, setPage] = useState(1)
-
-  const [selectedBoard, setSelectedBoard] = useState('')
-  const [search, setSearch] = useState('')
-
-  const [sortBy, setSortBy] = useState<SortOption>('latest')
-  const [checkedMap, setCheckedMap] = useState<Record<string, boolean>>({})
-
-  const sessionUser = useSessionStore((s) => s.user)
-  const isClient = useSyncExternalStore(
-    (onStoreChange) => {
-      queueMicrotask(onStoreChange)
-      return () => {}
-    },
-    () => true,
-    () => false
-  )
-
-  const user = isClient ? sessionUser : null
-
-  const myPostsQuery = useMyPosts(
-    { page, size: 10, sort: sortBy },
-    { enabled: isClient && tab === 'post' }
-  )
-  const myCommentsQuery = useMyComments(
-    { page, size: 15, sort: sortBy },
-    { enabled: isClient && tab === 'comment' }
-  )
-  const likedPostsQuery = useLikedPosts(
-    { page, size: 10, sort: sortBy },
-    { enabled: isClient && tab === 'like' }
-  )
-
-  const deleteMyPosts = useDeleteMyPosts()
-  const deleteMyComments = useDeleteMyComments()
-
-  const lastErrorKeyRef = useRef<string | null>(null)
-
-  const timelineHistoryQuery = useTimelineHistory(isClient)
-
-  const noHistoryToastOnceRef = useRef(false)
-  useEffect(() => {
-    if (!timelineHistoryQuery.isSuccess) return
-    if (noHistoryToastOnceRef.current) return
-
-    const results = timelineHistoryQuery.data?.results ?? []
-    if (results.length === 0) {
-      noHistoryToastOnceRef.current = true
-      toast('아직 출석 기록이 없습니다')
-    }
-  }, [timelineHistoryQuery.isSuccess, timelineHistoryQuery.data?.results])
-
-  const timeline = useMemo<TimelineItem[]>(() => {
-    const today = startOfKstDay()
-    const results = timelineHistoryQuery.data?.results ?? []
-
-    const submittedMap = new Map<string, boolean>()
-    for (const r of results) submittedMap.set(r.date, !!r.is_submitted)
-
-    const items: TimelineItem[] = []
-    for (let offset = -3; offset <= 3; offset += 1) {
-      const d = addDays(today, offset)
-      const ymd = toYmd(d)
-      const submitted = submittedMap.get(ymd) ?? false
-
-      let status: TimelineItem['status'] = 'upcoming'
-      if (offset > 0) status = 'upcoming'
-      else if (offset === 0) status = submitted ? 'done' : 'go'
-      else status = submitted ? 'done' : 'fail'
-
-      items.push({
-        date: toMmDd(d),
-        day: toKoreanDay(d),
-        status,
-      })
-    }
-
-    return items
-  }, [timelineHistoryQuery.data?.results])
-
-  const profile = useMemo(() => {
-    return {
-      nickname: user?.nickname ?? '',
-      email: user?.email ?? '',
-      joinedAt: '-',
-      profileImageSrc: user?.profileImageUrl ?? null,
-      balloonLeft: {
-        title: '오늘도 힘내봐요!',
-        subtitle: 'Hazlo lo mejor que puedas hoy también',
-      },
-      balloonRight: {
-        title: 'STUDY GO !',
-      },
-    }
-  }, [user?.email, user?.nickname, user?.profileImageUrl])
-
-  const posts = useMemo<MyPagePostItem[]>(() => {
-    const author = user?.nickname ?? ''
-    const avatar = user?.profileImageUrl
-    const safeAvatar = avatar ?? DEFAULT_AVATAR
-
-    return (myPostsQuery.data?.posts ?? []).map((p) => {
-      const { date, time } = formatDateParts(p.createdAt)
-      return {
-        id: p.id,
-        author,
-        date,
-        time,
-        title: p.title,
-        views: 0,
-        likes: 0,
-        comments: 0,
-        avatar: safeAvatar,
-        thumbnail: DEFAULT_THUMBNAIL,
-        board: 'free',
-      }
-    })
-  }, [myPostsQuery.data?.posts, user?.nickname, user?.profileImageUrl])
-
-  const likedPosts = useMemo<MyPagePostItem[]>(() => {
-    const author = user?.nickname ?? ''
-    const avatar = user?.profileImageUrl
-    const safeAvatar = avatar ?? DEFAULT_AVATAR
-
-    return (likedPostsQuery.data?.posts ?? []).map((p) => {
-      const createdAt = p.likedAt ?? p.createdAt ?? ''
-      const { date, time } = formatDateParts(createdAt)
-      return {
-        id: p.id,
-        author,
-        date,
-        time,
-        title: p.title,
-        views: 0,
-        likes: 0,
-        comments: 0,
-        avatar: safeAvatar,
-        thumbnail: DEFAULT_THUMBNAIL,
-        board: 'free',
-      }
-    })
-  }, [likedPostsQuery.data?.posts, user?.nickname, user?.profileImageUrl])
-
-  const comments = useMemo<MyCommentItem[]>(() => {
-    const safe = myCommentsQuery.data?.comments ?? []
-    return safe.map((c) => {
-      return {
-        commentId: String(c.id),
-        postId: c.postId == null ? null : String(c.postId),
-        postTitle: c.postTitle ?? null,
-        content: c.content ?? null,
-        createdAt: c.createdAt,
-        board: null,
-      }
-    })
-  }, [myCommentsQuery.data?.comments])
-
-  const filteredPosts = useMemo(() => {
-    const board = normalize(selectedBoard)
-    const q = normalize(search)
-
-    return posts.filter((p) => {
-      const okBoard = !board || normalize(p.board) === board
-      const okSearch =
-        !q || normalize(p.title).includes(q) || normalize(p.author).includes(q)
-      return okBoard && okSearch
-    })
-  }, [posts, selectedBoard, search])
-
-  const filteredLikes = useMemo(() => {
-    const board = normalize(selectedBoard)
-    const q = normalize(search)
-
-    return likedPosts.filter((p) => {
-      const okBoard = !board || normalize(p.board) === board
-      const okSearch =
-        !q || normalize(p.title).includes(q) || normalize(p.author).includes(q)
-      return okBoard && okSearch
-    })
-  }, [likedPosts, selectedBoard, search])
-
-  const filteredComments = useMemo(() => {
-    const board = normalize(selectedBoard)
-    const q = normalize(search)
-
-    return comments.filter((c) => {
-      const okBoard = !board || normalize(c.board) === board
-      const okSearch =
-        !q ||
-        normalize(c.postTitle).includes(q) ||
-        normalize(c.content).includes(q)
-      return okBoard && okSearch
-    })
-  }, [comments, selectedBoard, search])
-
-  useEffect(() => {
-    const isError =
-      (tab === 'post' && myPostsQuery.isError) ||
-      (tab === 'comment' && myCommentsQuery.isError) ||
-      (tab === 'like' && likedPostsQuery.isError)
-
-    if (!isError) {
-      lastErrorKeyRef.current = null
-      return
-    }
-
-    const error =
-      (tab === 'post' && myPostsQuery.error) ||
-      (tab === 'comment' && myCommentsQuery.error) ||
-      (tab === 'like' && likedPostsQuery.error)
-
-    const message =
-      error instanceof Error ? error.message : error ? String(error) : ''
-    const errorKey = `${tab}:${message}`
-
-    if (lastErrorKeyRef.current === errorKey) return
-    lastErrorKeyRef.current = errorKey
-
-    if (message) {
-      console.error('[MyPage] query error:', error)
-      toast.error(`마이페이지 데이터를 불러오지 못했습니다. (${message})`)
-    } else {
-      toast.error('마이페이지 데이터를 불러오지 못했습니다.')
-    }
-  }, [
+  const {
     tab,
-    myPostsQuery.error,
-    myPostsQuery.isError,
-    myCommentsQuery.error,
-    myCommentsQuery.isError,
-    likedPostsQuery.error,
-    likedPostsQuery.isError,
-  ])
+    page,
+    selectedBoard,
+    search,
+    sortBy,
+    checkedMap,
 
-  const handleChangeTab = (nextTab: TabType) => {
-    setTab(nextTab)
-    setPage(1)
-    setCheckedMap({})
-  }
+    user,
+    profile,
+    timeline,
 
-  const handleChangeSortBy = (next: SortOption) => {
-    setSortBy(next)
-    setPage(1)
-    setCheckedMap({})
-  }
+    filteredPosts,
+    filteredComments,
+    filteredLikes,
 
-  const handleChangeBoard = (value: string) => {
-    setSelectedBoard(value)
-    setPage(1)
-    setCheckedMap({})
-  }
+    totalPages,
+    actionLabel,
+    isCurrentTabLoading,
 
-  const handleChangeSearch = (value: string) => {
-    setSearch(value)
-    setPage(1)
-    setCheckedMap({})
-  }
-
-  const handleToggleOne = (id: string) => {
-    setCheckedMap((prev) => ({
-      ...prev,
-      [id]: !prev[id],
-    }))
-  }
-
-  const actionLabel = tab === 'like' ? '해지하기' : '삭제하기'
-
-  const handleClickAction = async () => {
-    const selectedIds = Object.entries(checkedMap)
-      .filter(([, v]) => v)
-      .map(([id]) => Number(id))
-      .filter((id) => Number.isFinite(id))
-
-    const selectedCount = selectedIds.length
-
-    if (selectedCount === 0) {
-      toast.error('선택된 항목이 없습니다.')
-      return
-    }
-
-    try {
-      if (tab === 'post') {
-        await deleteMyPosts.mutateAsync(selectedIds)
-        toast.success('선택한 게시글을 삭제했습니다.')
-        setCheckedMap({})
-        return
-      }
-
-      if (tab === 'comment') {
-        await deleteMyComments.mutateAsync(selectedIds)
-        toast.success('선택한 댓글을 삭제했습니다.')
-        setCheckedMap({})
-        return
-      }
-
-      toast.success('좋아요를 해지했습니다.')
-      setCheckedMap({})
-    } catch {
-      toast.error('요청 처리에 실패했습니다.')
-    }
-  }
-
-  const totalPages = useMemo(() => {
-    if (tab === 'post') return myPostsQuery.data?.pagination.totalPages ?? 1
-    if (tab === 'comment')
-      return myCommentsQuery.data?.pagination.totalPages ?? 1
-    if (tab === 'like') return likedPostsQuery.data?.pagination.totalPages ?? 1
-    return 1
-  }, [
-    tab,
-    myPostsQuery.data?.pagination.totalPages,
-    myCommentsQuery.data?.pagination.totalPages,
-    likedPostsQuery.data?.pagination.totalPages,
-  ])
-
-  const isCurrentTabLoading =
-    !isClient ||
-    (tab === 'post' && myPostsQuery.isLoading) ||
-    (tab === 'comment' && myCommentsQuery.isLoading) ||
-    (tab === 'like' && likedPostsQuery.isLoading)
+    setPage,
+    handleChangeTab,
+    handleChangeBoard,
+    handleChangeSearch,
+    handleChangeSortBy,
+    handleToggleOne,
+    handleClickAction,
+  } = useMyPageController()
 
   return (
     <div className="bg-brand-white min-h-screen">
